// libs/sales/src/lib/customers/models/customer.model.ts
export class Customer {
  id: number = 0;
  nombre: string = '';
  apellido: string = '';
  razonSocial: string = '';
  documento: string = '';
  tipoDocumento: string = '';
  telefono: string = '';
  email: string = '';
  direccion: string = '';
  ciudad: string = '';
  provincia: string = '';
  codigoPostal: string = '';
  activo: boolean = true;
}